// src/app/shared/models/dictionary.model.ts
export interface IDictionaryModel {
  updateDate: string;
  createUserId: string;
  code: string;
  typeName: string;
  icon: string;
  orderBy: string;
  shortNameEn: string;
  searchKeyword: string;
  nameRu: string;
  parentNameEn: string;
  typeNameRu: string;
  parentCode: string;
  id: string;
  nameAz: string;
  createDate: string;
  parentNameAz: string;
  typeNameAz: string;
  parentTypeId: string;
  updateUserId: string;
  hiddenStatus: string;
  active: string;
  nameEn: string;
  parentId: string;
  shortNameRu: string;
  parentName: string;
  typeNameEn: string;
  parentNameRu: string;
  name: string;
  typeId: string;
  shortNameAz: string;
  shortName: string;
}